requirements.txt
# Core Django Framework
Django==5.1.6
djangorestframework==3.15.2

# Django Extensions and Utilities
django-cors-headers==4.6.0
django-environ==0.11.2
django-colorfield==0.11.0
django-filter==24.2
whitenoise==6.8.2

# Database
psycopg2-binary==2.9.10

# Image Processing
Pillow==11.1.0

# PDF Generation and Document Processing
reportlab==4.2.5
weasyprint==62.3
cairocffi>=0.9.0

# Data Science and Machine Learning
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2
nltk==3.9.1

# AI and GPT Integration
google-genai>=0.2.0

# HTTP Requests
requests==2.32.3

# Environment and Configuration
python-decouple==3.8
python-dotenv==1.0.1

# Production Server
gunicorn==23.0.0

# Email Backend for Production
sendgrid==6.11.0
django-sendgrid-v5==1.2.3

# Additional utilities that might be needed
html2text==2024.2.26
